+++ src/experiments/alr_experiment.rs
@@ -4,13 +4,13 @@
 pub struct AlrExperimentSettings {
     pub pacing_factor: f64,
     pub max_paced_queue_time: i64,
-    pub alr_bandwidth_usage_percent: isize,
-    pub alr_start_budget_level_percent: isize,
-    pub alr_stop_budget_level_percent: isize,
+    pub alr_bandwidth_usage_percent: i64,
+    pub alr_start_budget_level_percent: i64,
+    pub alr_stop_budget_level_percent: i64,
     // Will be sent to the receive side for stats slicing.
     // Can be 0..6, because it's sent as a 3 bits value and there's also
     // reserved value to indicate absence of experiment.
-    pub group_id: isize,
+    pub group_id: i64,
 }
 
 impl Default for AlrExperimentSettings {
